Front-end and user handling setup
Created HTML structure for the homepage, login, and registration pages,
along with a java script to validate registration requirements. Applied
styles for individual subpages and a general style.

Designed separate templates for the footer and header across all
subpages.

Implemented an object for transferring data from the login form and
added annotations for validating this data.

Created UserService and its implementations to handle user
registration.

Established controllers for the homepage, registration, and login
pages.