Add Game Table Handling
Introduced WebSocketService with its implementation for WebSocket
protocol handling.

Added new functions to the Game Model and Game Table Model.

Implemented a countdown for the game start.

Created an data transfer object for action-related with player slots.

Established TableService interface and its implementations for handling
actions related to the game table, along with a controller for the game
table.

Fixed Lobby Manager.