Game Table Frontend and Backend Enhancements
Frontend:
Added stomp.js library for WebSocket handling on the user's side.
Created HTML structure and styles for the game table and implemented JavaScript to enable various game-related actions:
- Taking/Leaving a game slot,
- Starting the game,
- Moving cards,
- Replace jokers,
- Throwing cards,
- Laying down cards,
- other utils.

Backend:
Created DTOs for each in-game action. Established GameService and its implementations to handle and validate player actions, along with a controller for the game.
Added new functionalities to the game table.