Add Game End
Implemented checking if the game can be concluded and determining the
winner. Reorganized and cleaned up the game service, its
implementations, and the controller.
Added the endgame method in the GameTable class to conclude the game.
Reset Game Collections in the Game class constructor.
Created EndGameDto for data transfer between the server and the user.
Reorganized JS code, created functions responsible for handling game
termination.